TDP-2802: fix refresh import types problem
