+++ dataprep-webapp/src/app/services/import/import-service.js
@@ -15,23 +15,41 @@
  */
 export default class ImportService {
 
-	constructor($document, $rootScope, $translate, appSettings, ImportRestService, StateService) {
+	constructor($document, $rootScope, $translate, appSettings, state, DatasetService, ImportRestService,
+				StateService, TalendConfirmService, UploadWorkflowService, UpdateWorkflowService, SettingsService) {
 		'ngInject';
 
-		this.$rootScope = $rootScope;
 		this.appSettings = appSettings;
+		this.$document = $document;
+		this.$rootScope = $rootScope;
+		this.$translate = $translate;
 
 		this.ImportRestService = ImportRestService;
 		this.StateService = StateService;
 
+		this.$document = $document;
+		this.$translate = $translate;
+		this.state = state;
+		this.DatasetService = DatasetService;
+		this.SettingsService = SettingsService;
+		this.StateService = StateService;
+		this.TalendConfirmService = TalendConfirmService;
+		this.UpdateWorkflowService = UpdateWorkflowService;
+		this.UploadWorkflowService = UploadWorkflowService;
 
-		//TODO
 		this.currentInputType = null;
 		this.datastoreFormActions = null;
 		this.datasetFormActions = null;
 		this.dataStoreId = null;
-		this.$document = $document;
-		this.$translate = $translate;
+		this.updateDatasetFile = null;
+		this.importDatasetFile = null;
+		this.importTypes = this.state.import.importTypes;
+
+		this.onDatastoreFormChange = this.onDatastoreFormChange.bind(this);
+		this.onDatastoreFormSubmit = this.onDatastoreFormSubmit.bind(this);
+
+		this.onDatasetFormChange = this.onDatasetFormChange.bind(this);
+		this.onDatasetFormSubmit = this.onDatasetFormSubmit.bind(this);
 	}
 
 	manageLoader(method, args) {
@@ -256,5 +274,207 @@ export default class ImportService {
 		this.datastoreForm = null;
 		this.dataStoreId = null;
 		this.datasetForm = null;
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name onFileChange
+	 * @methodOf data-prep.services.import.service:ImportService
+	 * @description Update dataset
+	 */
+	onFileChange() {
+		this.UpdateWorkflowService.updateDataset(this.updateDatasetFile[0], this.state.inventory.datasetToUpdate);
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name onDatastoreFormChange
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @description Datastore form change handler
+	 * @param formData All data as form properties
+	 * @param formId ID attached to the form
+	 * @param propertyName Property which has triggered change handler
+	 */
+	onDatastoreFormChange(formData, formId, propertyName) {
+		const definitionName = formId || this.currentInputType.locationType;
+		this.refreshParameters(definitionName, propertyName, formData)
+			.then((response) => {
+				this.datastoreForm = response.data;
+			});
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name onDatastoreFormSubmit
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @description Datastore form change handler
+	 * @param uiSpecs All data as form properties
+	 * @param formId ID attached to the form
+	 */
+	onDatastoreFormSubmit(uiSpecs, formId) {
+		const definitionName = formId || this.currentInputType.locationType;
+		this.testConnection(definitionName, uiSpecs && uiSpecs.formData)
+			.then((response) => {
+				this.dataStoreId = response.data && response.data.dataStoreId;
+				if (!this.dataStoreId) {
+					return null;
+				}
+				return this.getDatasetForm(this.dataStoreId);
+			})
+			.then((datasetFormResponse) => {
+				this.datasetForm = datasetFormResponse && datasetFormResponse.data;
+			});
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name onDatasetFormChange
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @description Datastore form change handler
+	 * @param formData All data as form properties
+	 * @param formId ID attached to the form
+	 * @param propertyName Property which has triggered change handler
+	 */
+	onDatasetFormChange(formData, formId, propertyName) {
+		this.refreshDatasetForm(this.dataStoreId, propertyName, formData)
+			.then((response) => {
+				this.datasetForm = response.data;
+			});
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name onDatasetFormSubmit
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @description Datastore form change handler
+	 * @param uiSpecs
+	 */
+	onDatasetFormSubmit(uiSpecs) {
+		this.createDataset(this.dataStoreId, uiSpecs && uiSpecs.formData)
+			.then((response) => {
+				const dataSetId = response.data && response.data.dataSetId;
+				this.DatasetService.getDatasetById(dataSetId).then(this.UploadWorkflowService.openDataset);
+			});
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name importDataset
+	 * @description Create dataset using import parameters
+	 * @param {object} file The file imported from local
+	 * @param {string} name The dataset name
+	 * @param {object} importType The import parameters
+	 */
+	importDataset(file, name, importType) {
+		const params = this.DatasetService.getLocationParamIteration({}, importType.parameters);
+		params.type = importType.locationType;
+		params.name = name;
+
+		const dataset = this.DatasetService.createDatasetInfo(file, name);
+		this.StateService.startUploadingDataset(dataset);
+
+		return this.DatasetService.create(params, importType.contentType, file)
+			.progress((event) => {
+				dataset.progress = parseInt((100.0 * event.loaded) / event.total, 10);
+			})
+			.then((event) => {
+				this.DatasetService.getDatasetById(event.data).then(this.UploadWorkflowService.openDataset);
+			})
+			.catch(() => {
+				dataset.error = true;
+			})
+			.finally(() => {
+				this.StateService.finishUploadingDataset(dataset);
+			});
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name import
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @description Import step 1 - It checks if the dataset name is available
+	 * If so : the dataset is created
+	 * If not : the new name modal is shown
+	 */
+	import(importType) {
+		const file = this.importDatasetFile ? this.importDatasetFile[0] : null;
+		const datasetName = file ?
+			file.name :
+			_.find(importType.parameters, { name: 'name' }).value;
+
+		// remove file extension and ask final name
+		const name = datasetName.replace(/\.[^/.]+$/, '');
+
+		return this.DatasetService.checkNameAvailability(name)
+			// name available: we create the dataset
+			.then(() => {
+				this.importDataset(file, name, importType);
+			})
+			// name is not available, we ask for a new name
+			.catch(() => {
+				this.datasetName = name;
+				this.datasetNameModal = true;
+			})
+			.finally(() => {
+				this.StateService.setShowImportModal(false);
+			});
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name uploadDatasetName
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @description Import step 2 - name entered. It checks if the name is available
+	 * If so : the dataset is created
+	 * If not : the user has to choose to create a new one or the update the existing one
+	 */
+	onImportNameValidation() {
+		const file = this.importDatasetFile ? this.importDatasetFile[0] : null;
+		const importType = this.currentInputType;
+		const name = this.datasetName;
+
+		return this.DatasetService.checkNameAvailability(name)
+			// name still exists
+			.then(() => {
+				this.importDataset(file, name, importType);
+			})
+			// name still exists : we ask if user want to update it
+			.catch(existingDataset => this.updateOrCreate(file, existingDataset, importType, name));
+	}
+
+	/**
+	 * @ngdoc method
+	 * @name updateOrCreate
+	 * @methodOf data-prep.import.controller:ImportCtrl
+	 * @param {object} file The dataset file
+	 * @param {object} existingDataset The dataset to update
+	 * @param {object} importType The import configuration
+	 * @param {string} name The dataset name
+	 * @description Import step 3 - Ask to create or update the existing dataset
+	 * Create : get a unique name and create
+	 * Update : update the content of the existing dataset
+	 */
+	updateOrCreate(file, existingDataset, importType, name) {
+		return this.TalendConfirmService.confirm(null, ['UPDATE_EXISTING_DATASET'], { dataset: name })
+			// user confirm : let's update the dataset
+			.then(() => {
+				this.UpdateWorkflowService.updateDataset(file, existingDataset);
+			})
+			// user dismiss : cancel
+			// user select no : get unique name and create a new dataset
+			.catch((cause) => {
+				if (cause === 'dismiss') {
+					return;
+				}
+
+				return this.DatasetService.getUniqueName(name)
+					.then((name) => {
+						return this.importDataset(
+							file,
+							name,
+							importType
+						);
+					});
+			});
 	}
 }
